Modify test cases to not compile-on-list
Changes tests that inadvertently trigger the compiler when doing test
discovery/listing.
