+++ numba/tests/test_serialize.py
@@ -12,7 +12,6 @@ from numba.core.errors import TypingError
 from numba.tests.support import TestCase
 from numba.core.target_extension import resolve_dispatcher_from_str
 from numba.cloudpickle import dumps, loads
-from .serialize_usecases import *
 
 
 class TestDispatcherPickling(TestCase):
@@ -51,33 +50,40 @@ class TestDispatcherPickling(TestCase):
             check_result(new_func)
 
     def test_call_with_sig(self):
+        from .serialize_usecases import add_with_sig
         self.run_with_protocols(self.check_call, add_with_sig, 5, (1, 4))
         # Compilation has been disabled => float inputs will be coerced to int
         self.run_with_protocols(self.check_call, add_with_sig, 5, (1.2, 4.2))
 
     def test_call_without_sig(self):
+        from .serialize_usecases import add_without_sig
         self.run_with_protocols(self.check_call, add_without_sig, 5, (1, 4))
         self.run_with_protocols(self.check_call, add_without_sig, 5.5, (1.2, 4.3))
         # Object mode is enabled
         self.run_with_protocols(self.check_call, add_without_sig, "abc", ("a", "bc"))
 
     def test_call_nopython(self):
+        from .serialize_usecases import add_nopython
         self.run_with_protocols(self.check_call, add_nopython, 5.5, (1.2, 4.3))
         # Object mode is disabled
         self.run_with_protocols(self.check_call, add_nopython, TypingError, (object(), object()))
 
     def test_call_nopython_fail(self):
+        from .serialize_usecases import add_nopython_fail
         # Compilation fails
         self.run_with_protocols(self.check_call, add_nopython_fail, TypingError, (1, 2))
 
     def test_call_objmode_with_global(self):
+        from .serialize_usecases import get_global_objmode
         self.run_with_protocols(self.check_call, get_global_objmode, 7.5, (2.5,))
 
     def test_call_closure(self):
+        from .serialize_usecases import closure
         inner = closure(1)
         self.run_with_protocols(self.check_call, inner, 6, (2, 3))
 
     def check_call_closure_with_globals(self, **jit_args):
+        from .serialize_usecases import closure_with_globals
         inner = closure_with_globals(3.0, **jit_args)
         self.run_with_protocols(self.check_call, inner, 7.0, (4.0,))
 
@@ -88,22 +94,27 @@ class TestDispatcherPickling(TestCase):
         self.check_call_closure_with_globals(forceobj=True)
 
     def test_call_closure_calling_other_function(self):
+        from .serialize_usecases import closure_calling_other_function
         inner = closure_calling_other_function(3.0)
         self.run_with_protocols(self.check_call, inner, 11.0, (4.0, 6.0))
 
     def test_call_closure_calling_other_closure(self):
+        from .serialize_usecases import closure_calling_other_closure
         inner = closure_calling_other_closure(3.0)
         self.run_with_protocols(self.check_call, inner, 8.0, (4.0,))
 
     def test_call_dyn_func(self):
+        from .serialize_usecases import dyn_func
         # Check serializing a dynamically-created function
         self.run_with_protocols(self.check_call, dyn_func, 36, (6,))
 
     def test_call_dyn_func_objmode(self):
+        from .serialize_usecases import dyn_func_objmode
         # Same with an object mode function
         self.run_with_protocols(self.check_call, dyn_func_objmode, 36, (6,))
 
     def test_renamed_module(self):
+        from .serialize_usecases import get_renamed_module
         # Issue #1559: using a renamed module (e.g. `import numpy as np`)
         # should not fail serializing
         expected = get_renamed_module(0.0)
@@ -111,6 +122,7 @@ class TestDispatcherPickling(TestCase):
                                 expected, (0.0,))
 
     def test_call_generated(self):
+        from .serialize_usecases import generated_add
         self.run_with_protocols(self.check_call, generated_add,
                                 46, (1, 2))
         self.run_with_protocols(self.check_call, generated_add,
@@ -121,6 +133,7 @@ class TestDispatcherPickling(TestCase):
         Check that reconstructing doesn't depend on resources already
         instantiated in the original process.
         """
+        from .serialize_usecases import closure_calling_other_closure
         func = closure_calling_other_closure(3.0)
         pickled = pickle.dumps(func)
         code = """if 1:
@@ -140,6 +153,7 @@ class TestDispatcherPickling(TestCase):
 
         Note that "same function" is intentionally under-specified.
         """
+        from .serialize_usecases import closure
         func = closure(5)
         pickled = pickle.dumps(func)
         func2 = closure(6)
@@ -207,7 +221,6 @@ class TestSerializationMisc(TestCase):
         got2 = _numba_unpickle(id(random_obj), bytebuf, hashed)
         # unpickled results are the same objects
         self.assertIs(got1, got2)
-
 
 
 class TestCloudPickleIssues(TestCase):
